apply feedback from orta and rubocop
