+++ lib/danger/comment_generators/github_inline.md.erb
@@ -1,6 +1,6 @@
 <%- @tables.each do |table| -%>
   <%- if table[:content].any? -%>
-<table>
+<table data-meta="generated_by_<%= @danger_id %>">
   <tbody>
     <%- table[:content].each do |violation| -%>
     <tr>
@@ -17,6 +17,10 @@
 <%= current %>
 <%# the previous line has to be aligned far to the left, otherwise markdown can break easily %>
 <%- end -%>
+<%# We need to add the generated_by_ to identify danger origined comments. But with inlines %>
+<%# it might be a little annoying, so we set on the table, but if we have markdown we add the footer anyway %>
+<%- if @markdowns.count > 0 -%>
 <p align="right" data-meta="generated_by_<%= @danger_id %>">
   Generated by :no_entry_sign: <a href="http://danger.systems/">danger</a>
 </p>
+<%- end -%>
